# docs/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: AI Enhanced PDF Scholar API
  description: |
    **Enterprise-grade intelligent document management platform for academic research**
    
    This API provides comprehensive functionality for:
    - 🎯 **Intelligent Document Processing**: Smart import with metadata extraction and duplicate detection
    - 📊 **Advanced Citation Analysis**: Multi-format parsing with network discovery
    - 🧠 **RAG (Retrieval-Augmented Generation)**: AI-powered document queries and analysis
    - 🔐 **Enterprise Security**: JWT authentication with RBAC and XSS protection
    - ⚡ **Real-time Features**: WebSocket support for streaming operations
    - 📈 **Performance Monitoring**: APM with metrics and health checks
    
    ## Authentication
    
    Most endpoints require JWT authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    
    API endpoints are rate-limited to ensure fair usage:
    - **Default**: 100 requests per minute per IP
    - **Global IP**: 1000 requests per hour per IP
    - **Authenticated users**: Higher limits apply
    
    ## Error Handling
    
    The API uses standard HTTP status codes and provides detailed error information:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource not found)
...
    
  version: "2.1.0"
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/rag/query:
    post:
      tags: [rag]
      summary: Query documents with AI
      description: |
        Query documents using RAG (Retrieval-Augmented Generation) for intelligent answers.
        Supports context-aware queries with semantic search and citation tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Natural language query
                  example: "What are the main findings about transformer architectures?"
                document_id:
                  type: integer
                  description: Specific document to query (optional)
                context_window:
                  type: integer
                  default: 4000
                  description: Context window size for query processing
                include_sources:
                  type: boolean
                  default: true
                  description: Include source citations in response
                temperature:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.1
                  description: Response creativity (0=factual, 1=creative)
      responses:
        '200':
          description: RAG query completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGQueryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Document not found or not indexed
  /api/rag/index/{document_id}:
    post:
      tags: [rag]
      summary: Index document for RAG
      description: Create or rebuild vector index for document to enable RAG queries
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                chunk_size:
                  type: integer
                  default: 1000
                  description: Size of text chunks for indexing
                chunk_overlap:
                  type: integer
                  default: 200
                  description: Overlap between chunks
                force_rebuild:
                  type: boolean
                  default: false
                  description: Force rebuild existing index
      responses:
        '200':
          description: Document indexed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexingResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          description: Document format not supported for indexing

  # Citation Analysis Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: integer
      description: Unique document identifier
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
            error_code: "NOT_FOUND"
  schemas:
...
    BaseResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error_code:
              type: string
              description: Machine-readable error code
            details:
              type: object
              description: Additional error details
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            validation_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
...
    RAGQueryResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                query:
                  type: string
                  description: Original query
                answer:
                  type: string
                  description: AI-generated answer
                sources:
                  type: array
                  items:
                    type: object
                    properties:
                      document_id:
                        type: integer
                      document_title:
                        type: string
                      page_number:
                        type: integer
                      confidence:
                        type: number
                      excerpt:
                        type: string
                processing_time:
                  type: number
                  description: Query processing time in seconds
                context_tokens:
                  type: integer
                  description: Number of tokens used for context
    IndexingResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                document_id:
                  type: integer
                chunks_created:
                  type: integer
                index_size:
                  type: integer
                  description: Index size in bytes
                processing_time:
                  type: number

    # Citation Models
...